fix: eliminate build warnings on Linux/Windows/macOS

+++ src/compio.cpp
@@ -2085,7 +2085,7 @@ uint64_t compio_erase(uint64_t size, compio_file *file) {
 
         const auto left_val = archive->index->get(left_to_merge.value());
         if (!left_val.has_value()) {
-            WARNING_PRINT("warning: left_to_merge key ({%lu, %lu}) was not found in the tree after "
+            WARNING_PRINT("warning: left_to_merge key ({%" PRIu64 ", %" PRIu64 "}) was not found in the tree after "
                           "erase operation\n",
                           left_to_merge->hash, left_to_merge->pos);
             goto after_merge;
@@ -2093,14 +2093,14 @@ uint64_t compio_erase(uint64_t size, compio_file *file) {
         const auto right_val = archive->index->get(right_to_merge.value());
         if (!right_val.has_value()) {
             WARNING_PRINT(
-                "warning: right_to_merge key ({%lu, %lu}) was not found in the tree after "
+                "warning: right_to_merge key ({%" PRIu64 ", %" PRIu64 "}) was not found in the tree after "
                 "erase operation\n",
                 right_to_merge->hash, right_to_merge->pos);
             goto after_merge;
         }
 
-        const auto block_size__minimum = archive->config.block_size__minimum;
-        const auto block_size__maximum = archive->config.block_size__maximum;
+        const auto block_size__minimum = static_cast<uint64_t>(archive->config.block_size__minimum);
+        const auto block_size__maximum = static_cast<uint64_t>(archive->config.block_size__maximum);
         const auto left_size = left_val->size;
         const auto right_size = right_val->size;
         DEBUG_PRINT("[CE]---postmerge sizes: %lu and %lu\n", left_size, right_size);
